Update patch set 1
Patch Set 1:

This is a feature to satisfy bugzilla https://bugs.linaro.org/show_bug.cgi?id=1166 "Make running GDB tests in Jenkins optional" which we identified as a blocker for moving Master to Stable.  Do we want to evaluate it for staging after we've proven that the existing staging produces no regressions vs. MergeFarm?

None-the-less, are there other patches waiting to go into Master?  If not staging can just be branched from the commit-ID previous to where this goes in.  If so, this can wait a bit but then a patch to add make_check return value tests should go in before staging is branched.

Patch-set: 1