Update patch set 1
Patch Set 1:

(4 comments)

All of the major changes asked for in the review have been completed by subsequent patches:

https://review.linaro.org/#/c/5064/
https://review.linaro.org/5065

Rob, you can review this now and consider it for inclusion.  I would suggest that you test the result of merging all three patch-sets for completeness.

There is a potential bug due to not quoting $1 in the first patch that's fixed in the others.

Patch-set: 1