Update patch set 1
Patch Set 1:

I agree with Ryan, the current behavior is OK.
And we can propose an enhancement later.

For the time being, we want to use the feature discussed here in our validation scripts, so adding the 2 options will be a one time thing. Having them bundled does not save us anything.

Patch-set: 1