Update patch set 1
Patch Set 1:

Just looking at generating a warning instead of an error: the easy way is to replace AC_MSG_ERROR by AC_MSG_WARN if $missing is not empty, but some of the packages which can be in $missing should really cause an error if not present (for instance: bash, git-new-workdir, wget, gcc)

I'm not sure which exact subset should cause an error, and which should remain a warning.

We can also leave the code as-is, forcing us to fix build-08.

Patch-set: 1
Reviewer: Christophe Lyon <1000072@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0