Update patch set 1
Patch Set 1:

This patch is a proposal to have a 'checkgdb' check box in Jenkins, to enable/disable gdb testsuite.

I wrote before Rob added support for --check, which does not enable to ask for all-but-gdb. However, I realize now that it might be better if I added support for --excludecheck, which would be more powerful than this.

Thoughts?

Patch-set: 1