Update patch set 1
Patch Set 1:

An alternative to abandoning in the future is that you do:

git review --list

Find this patch-set in the list (5340 in this case) and pull it down into your tree without committing the fix:

git review -N 5340
<resolve any conflicts>
<make your changes>
git commit

The git commit will reuse the commit message from review 5340 and will udpate the date.  When you then resubmit this fix using git review, it will 'update' the review in gerrit automatically.

Patch-set: 1