Update patch set 3
Patch Set 3:

My understanding is that it is the purpose of bug #1166.

I have added an excludecheck text field in Backport-staging and MergeFarm-staging.
It defaults to 'gdb' such that gdb testing will now be disabled by default for both cross and native configurations in Jenkins.

Leaving that field empty in Jenkins will enable gdb testing for both native and cross configurations. However, there is still an early exit in ABE's "make_check gdb" for cross configurations, for the time being.

Compared to the previous status, this patch removes the distinction between cross and native builds in jenkins.sh and test-backport.sh.

Patch-set: 3