Update patch set 3
Patch Set 3:

I don't think we can control that in Jenkins.

We describe a matrix of configurations, see for instance:
http://jenkins.tcwglab.linaro.org/job/Backport-staging/configure
(section "Configuration matrix")
on top of which we apply a "Combination filter" which decides which subset of the matrix is actually used.

But all the configurations are executed using the very same command, with some values in the environment (e.g. target, gcc_branch, etc....).

Patch-set: 3
Reviewer: Christophe Lyon <1000072@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0