Update patch set 1
Patch Set 1:

Yes, abe_top appears to be the dir-where-configure-is-run. My impression is that builds and snapshots are necessarily underneath that directory - please tell me if I'm wrong.

This patch just makes the existing size check not tied to how the machines we are using now happen to be set up. If we really want to do the size check properly, we not only need to consider --tarbin, but also cases where we build a subset of the toolchain, or something completely different, such as a small benchmark. Absent that, we might want to make the size check overrideable.

Patch-set: 1
Reviewer: Bernie Ogden <1000103@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0