Rework handling of reference snapshots used by Jenkins jobs
This patch fixes various problems that manifest themselves as "SHA1 not found"
in the repository.

Reference snapshots are used to speed-up "git clone" of repositories and
"wget URL" of tarball components.  The reference repositories are shared
among all builds and care needs to be taken to access them in read-only mode.
Specifically, the only safe mode of access to the reference git repos is via
"git clone --reference REFERENCE_REPO http://upstream_repo local_repo".

On the other hand, we want to update reference snapshots from time-to-time
(using a provided script) and those updates need to be coordinated with
"git clone"s that ABE does.  For this reason we separate checkout and build/test
stages of ABE, and protect checkout stage with a lock.  Reference snapshot
updates grab this lock in exclusive mode for the updates, while ABE builds grab
this lock in shared mode when doing source checkout.

Additionally, this patch adds logic to lib/fetch.sh to copy infrastructure/*
tarballs from reference copy, instead of always downloading them from
a web-site.  Previously, only Git repo were using reference snapshots.

As a future improvement, we should rename --with-git-reference-dir= ABE
configure option to --with-reference-snapshots= to better correspond to its
current semantics.

Change-Id: I3334a032fa0b422891b708bb258a0e1adf9d81b2