Update patch set 1
Patch Set 1:

The main purpose of this patch is to add a --mcpu option so I can build toolchains that default to a particular CPU, to test support for that CPU.  I noticed that --march was broken, and --mtune is missing, so I fixed that also.

There is some unfortunate duplication in the gcc.conf file in the arm targets.  I haven't come up with a good idea on how to avoid that yet.

I built about a dozen toolchains to test various option combinations with various targets, and it seems to be working as expected.  gcc-4.9 aarch64 doesn't support any of these options, you have to use the master branch (aka pre-release 5.0).

Patch-set: 1