Update patch set 1
Patch Set 1:

It should be noted that this change is orthogonal to any manifest rework that Rob is doing.  This patch corrects uses of reference snapshot directory.  The patch makes (1) ABE not modify reference snapshots and (2) checkout sources before building them (as opposed to during the build).  The checkout is done standalone so that use of reference snapshots is localized and protected by a lock.

Rob,

Would you please review this version, and I will then merge a rebased version for master.  [I initially implemented this against master, but could not test it due to master failing for unrelated reasons.]

Thank you.

Patch-set: 1