Update patch set 1
Patch Set 1: Code-Review-1

I'm curious what your problems were with master, it's working fine for me. Anyway, I refer to have the checkout happen during the build as the primary way of getting sources. This is so somebody can builds a single component with different sources in an existing build.

Patch-set: 1