Update patch set 1
Patch Set 1: -Code-Review

When using --checkout all, that's fine, as it's under user control. When not using --checkout, all sources should be downloaded or fetched during the builds process. As long as the checkout in bulk, or during the build is controllable by the user, I have less issues.

Patch-set: 1