Update patch set 1
Patch Set 1:

The docs should not recommend using -none- when configuring for linux targets, because that is not how we build the toolchains we use internally and release to customers.  This just confuses people new to ABE who take the time to read the docs.

Also, there are at least 3 places that fail if you configure a -none- linux target, because they have code that assumes that only bare targets have -none- in the tuple.

Patch-set: 1
Label: Code-Review=0