${force} now overwrites existing files in fetch_http()
The wget program will not overwrite a file unless -O is used to specify the
output file name on the command line.  This patch enables -O
${local_snapshots}/${getfile} to overwrite the existing file if ${force} is
specified.

It also moves a check for ${force} to the top of the function
so that fetch_http() won't return 0 automatically if ${force} is specified
but the ${getfile} already exists.  This also allows us to squash an
if-block around the wget call.

A test-case was added to verify that fetch_http works with ${force} = yes.

Change-Id: I685011b3383bd0aa7f1cf01a31ca22b77c0c5d41