Update patch set 2
Patch Set 2:

In testing patch-set one I discovered that a fix was needed in how wget was invoked.  That has resulted in a now dependent patch-set: https://review.linaro.org/5948

Christophe identified another bug in patch-set 1 which I've fixed in this series where the 'return 1' was in the wrong place.

I suspect that the automated testing fails because of the dependency on https://review.linaro.org/5948

I added new test-cases which catch the error conditions and verify that they're fixed.

Additionally, other problems were discovered in fetch_http() with ${force} and ${wget_bin}.  I've created additional fixes that now depend on this fix:

https://review.linaro.org/5949
https://review.linaro.org/5950

Patch-set: 2
Label: Code-Review=0