Update patch set 1
Patch Set 1:

Rob, In the future please allow for a +1 before giving a +2.

To answer Christophe's questions, GCC appears to only install libraries into lib/ and if it uses lib64/ it's only for it's internal library paths and the compiler knows how to find those libraries.  Rob and I both did investigations on the resultant library paths and I didn't see anything in the sysroot.

I too have had concerns about the binary toolchain installation image paths, but my understanding is that conforming applications shouldn't depend on the hardcoded sysroot library paths.  Unfortunately Open Embedded is one of the worst offenders in assuming that the libraries are in a statically defined place.

There should be no problem with mixing sysroots, as the sysroots prepend the target directory and the target defines the bit-ness, i.e.,

$sysroot/aarch64-linux-gnu/lib/
$sysroot/arm-linux-gnueabihf/lib/

These can both exist in the same sysroot directory and there is no collision.

Patch-set: 1
Reviewer: Ryan Arnold <1000055@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0