Update patch set 1
Patch Set 1:

Ahh, this is interesting, especially considering that I ran this exact test in isolation on my laptop before submitting the change.  I wonder why these have failed.  If we run the test suite with verbosity turned on then we could see why it failed, but that would bloat the logs.

18:03:19 FAIL: 'VALID --dryrun --force --target arm-linux-gnueabihf --disable make_docs --build all'
18:03:20 FAIL: 'VALID --dryrun --force --target arm-linux-gnueabihf --enable make_docs --build all'
18:03:20 PASS: 'VALID --dump'
18:03:20 PASS: 'VALID --dump --disable install'
18:03:20 PASS: 'VALID --dump'
18:03:20 PASS: 'VALID --dump --enable install'
18:03:20 FAIL: 'VALID --dryrun --force --target arm-linux-gnueabihf --disable install --build all'
18:03:21 FAIL: 'VALID --dryrun --force --target arm-linux-gnueabihf --enable install --build all'

Patch-set: 1