Update patch set 6
Patch Set 6:

I think there is a problem with this patch that should be fixed:

Calling jenkins.sh with --snapshots ="${foo}" where there is an erroneous variable/expansion might blow away the wrong directory with the: rm -rf ${user_snapshots}.  I'm not sure execution would actually get this far in such a case but I didn't want to test it to find out.

I'm not sure if there's a good way to sanity check this.

Patch-set: 6
Label: Code-Review=0