Update patch set 1
Patch Set 1: Code-Review+2

I'm going to approve this patch because it's clean and solves an immediate problem but really this patch is overloading the purpose of fetch_http().

To really do this cleanly I think fetch() and fetch_http() need to be reworked.

The md5sums fetching code should be broken into a separate fetch_md5sums() function which is called from the top-level abe.sh script and all of that code removed from fetch() proper.  This is because it's only called from one place and called early and it's already special-cased in fetch(). 

Much of what happens in fetch_http() regarding looking at snapshots/ for an existing tarball should be done in fetch().  I think fetch_http should simply fetch using http unconditionally if we already know we need to do it.

A new function, called fetch_reference_snapshot() would be called instead of fetch_http() if we are using a reference-snapshot dir.

Additionally fetch_http() should have wget use the server timestamps so it only downloads if the server version is newer than the local.  This would save on bandwidth.

Patch-set: 1
Reviewer: Ryan Arnold <1000055@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=+2
Label: SUBM=+1