Update patch set 1
Patch Set 1: Code-Review+1

(5 comments)

I added some comments as well.  There are some strange white-space issues.  There's also an issue with modifying some 'manifest' issues with this patch which should be in a different patch.

Once the issues in the comments are addressed I think this can go in.

Patch-set: 1