Update patch set 6
Patch Set 6:

I'm not very happy with changing the current logic: let's say you commit the ABE patch to disable ALSR when using QEMU for ARMEB. How can I re-run the same validation build, and compare with the previous one to make sure the patch works as expected?

Even if I start a buildfarm job with a specific logserver, it doesn't seem that I'll be able to compare 2 builds with different logservers. Or, is $logserver part of ref_log/newlog?

I think I'd rather keep jenkins.sh as-is and possibly have another script  handle the locking/dependencies

Patch-set: 6
Label: Code-Review=0