scripts/clean-schroot-sessions.sh: New script to clean stale schroot sessions
This script supports tcwg-cleanup-tcwg-test-schroots jenkins job, which cleans
up stale schroots daily.  The reason for stale schroot sessions to appear is
in how jenkins handles ssh agent.

If a build is cancelled, ssh agent dies immediatelly, and the builder can no
longer access the tester to finish the session (and it tries, as part of
"trap" cleanup).  Therefore, once a day, we stop schroot sessions of
tcwg-buildslave that lasted for more than a day.

Change-Id: Ib47f7af05ce22b07ed060a09e4a03cc19ed7cbfd