Early exit in case the git clone command used by jenkins.sh fails.
If, for some reason, the 'git clone' command used by jenkins at the beginning of the build fails, the build would continue until it reaches GCC build. It would fail at that point for lack of actual sources, with a possibly confusing message.
It's better to exit as soon as we know there was a problem.

Change-Id: If43e2b5f710cdcbac4bd58fcaf89016caeaceda4