scripts/compare_jobs.sh: Use specific log directories
In current work-in-progress backport job results are uploaded to specific
directories on $logserver, to avoid any ambiguity in their location.
This patch updates compare_jobs.sh script to not use hard-coded log paths.

This patch does two things: firstly, it changes default template for log
directories to only use variables, that can be calculated from input arguments.
Secondly, it moves calculation of those variables to the beginning of
the script, so that $logname can be computed, and checked for existence, and
the build/test can be, potentially, skipped altogether.

Note, this change is, mostly, for BuildFarm-style jobs, as Backport-style jobs
override $logname's of their children BuildFarm jobs to a known value to avoid
any guesswork.

Change-Id: Ib6a0dd141ac35e5b2bb2fbd9e10aa2edeeef43d8