Update patch set 1
Patch Set 1:

Configuring for --host i686-w64-mingw32 gives an error that doesn't stop the build.
/home/wilson/Linaro/tcwg/abe/config/gdb.conf: command substitution: line 17: unexpected EOF while looking for matching `''
/home/wilson/Linaro/tcwg/abe/config/gdb.conf: command substitution: line 18: syntax error: unexpected end of file

There is a generic problem here that if you include a X.conf file and get an error, it is ignored.  I'm not trying to fix that here.

The specific problem fixed by this patch is that the mingw32 sed command is missing a close quote.  With that fixed, the build now fails because the sed command failed to match a space that doesn't exist after tui.  With the space removed from the sed pattern the build works properly.

There is a similar problem in gcc.conf found by observation.  It is commented out, but should still be fixed, and needs similar fixes.

Patch-set: 1
Reviewer: Jim Wilson <1000132@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0