Fix problem with jeknins.sh locks hanging forever
Note on '-tt': We forcefully allocate pseudo-tty for the flock command
so that flock dies (through SIGHUP) and releases the lock when this
script is cancelled or terminated for whatever reason.  Without SIGHUP
reaching flock we risk situations when a lock will hang forever preventing
any subsequent builds to progress.  There are a couple of options as to
exactly how enable delivery of SIGHUP (e.g., set +m), and 'ssh -tt' seems
like the simplest one.

Change-Id: Ie7b3148b7397020d2ebf3b4af7eecc462ade5c70