Update patch set 1
Patch Set 1: Code-Review+1

I "+1" on the quality of the code, but I don't quite understand motivation for this patch.  The commit log positions this as a fix, but it is not clear what the problem is.

If this is a fix for a performance problem that scanning of md5sums does a lot of needless work, then please consider whether performance benefits worth the added code.  If this is a fix for a behavior problem (possibly exposed by a later patch), then this is OK.

Patch-set: 1
Reviewer: Maxim Kuvyrkov <1000092@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=+1