Update patch set 1
Patch Set 1:

I had to put a warning someplace as documentation. The proper fix is to have an installer for win32 releases that sets these two environment variables in the Windows registry, I'm working on that, but it'll be a separate patch.

Patch-set: 1
Reviewer: Rob Savoye <1000040@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0