Add configure tests for alien and fakeroot.
Alien and fakeroot are required only to build binary DEB packages,
which is optional.

Installing Alien installs 'dpkg' so we need to modify the test that
determines whether we're on DEB or RPM based systems to only discover
a DEB based system if dpkg is installed and pacman/yum is not.

Regenerate configure from configure.ac.

Change-Id: I6ba3fb1a2f62518276bf1cce1a2386c32fd982f7