Update patch set 1
Patch Set 1:

> Is dnf functionally equivalent on Fedora20 to yum?  I'm still stuck
 > on Fedora20.

I don't have an F20 install to verify it, but dnf has been in Fedora since 18 and I'll be surprised if the most basic functions (makecache and install) don't work.  It has been touted as a full replacement since F20 and has actually moved in as the primary command in F21 with yum merely an alias to it with an ugly warning.

Patch-set: 1
Reviewer: Siddhesh Poyarekar <1000178@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0