Update patch set 1
Patch Set 1:

This change has been tested here:
https://ci.linaro.org/view/tcwg-ci/job/tcwg-compare-results-master/444/

It works, but we'll have to update the Jenkins tcwg-compare-results-master.yaml job, to stop trying to copy ./scripts/unstable-tests.txt. This causes an error message in the console since the file is removed by this patch, but the error is harmless since the list is downloaded from another repo.

Patch-set: 1