Update patch set 1
Patch Set 1:

An error message appears in the logs:
ERROR: cpu0.gicv3_cpu_if: Virtual LPIs can not be supported in GICv2 only mode. Defaulting virtual lpi support to false.

for each execution, but it seems harmless.

I did verify manually and the code is actually executed and the return code is 0.

Patch-set: 1