Update patch set 1
Patch Set 1:

Before rejecting a patch from your fellow teammate with a -1, it is a good idea to ask why the patch is required, if that it is not obvious from the commit message.

In the case of this change, the patch is needed to "Fix building libgloss in bare-metal configs".  ABE:test branch currently builds only newlib, and misses libgloss, which results in a broken bare-metal toolchain.  Since it is ABE's goal to build toolchains with working newlib and present libgloss, I think this patch is necessary.

Patch-set: 1