Update patch set 2
Patch Set 2: Code-Review+1

(2 comments)

Rob -- OK with the inline comments addressed.  My preference is to maintain this script outside of ABE, since we may want to iterate on it quicker than on the ABE itself.  E.g., for production we will use abe:stable branch, but for testing the release it's OK to use a master branch of another repo.  Please consider moving this script to, say, git.linaro.org/toolchain/release-testing-scripts.git or somesuch.

Christophe -- regarding your comment on removing the $testdir -- I think it's fine to support non-clean environment in the scripts, since we depend on jenkins jobs to wipe and clean the workspace.

Patch-set: 2
Label: Code-Review=+1