Update patch set 2
Patch Set 2:

Bernie,

The failure you are seeing regarding master.conf seems to be an unrelated bug in ABE in handling binutils branches.  Since both binutils and GDB are in the same repository, ABE uses branch names to figure out the component the setting is addressed to.

There seems to be a bug in handling of binutils=binutils-gdb.git~master (would you please file it in bugzilla?) that causes ABE to think "master" is a new toolchain component.  As a workaround, just remove binutils customization from your jenkins job and use the ABE default.

Patch-set: 2