Update patch set 1
Patch Set 1:

on https://review.linaro.org/#/c/10687/1 Maxim wrote:

> GDB and Binutils are now using default make flags (see
> https://review.linaro.org/#/c/10799/).  Do we need
> anything else to commit this patch?

gdb is still using a non-default make target for gdbserver, and this results in errors with current abe master, so the build fails with this patch.

In addition, the non-default make targets are used for make_docs, which is also a possible source of fragility, although AFAICS it works at the moment.

Patch-set: 1