Update patch set 1
Patch Set 1:

(18 comments)

I've found a number of issues. I've -1'd this because of the issues which I've tagged as bugs.

There's two marked as [unsure]. A more accurate comment might be all that's needed.

The other issues which are tagged as [cleanup] could be left until after the merge is complete. Let me know which ones you don't want to fix immediately and I'll add them to bug #2165.

Patch-set: 1
Label: Code-Review=0