Look for manifest variables at start of line only.
When building in jenkins, jenkins.sh adds a build_url line to the manifest.

An example such line is:
   build_url=https://ci.linaro.org/jenkins/job/tcwg-bkk16-buildfarm/label=tcwg-x86_64-cam,target=native/115/

As the url includes the string 'target=', it causes the manifest parser to
incorporate parts of the url in the target name, causing a cascade of error
messages. This is more work to be done to improve the robustness of
import_manifest() but this patch addresses the immediate problem.

Change-Id: Ie14d555b8f4d2ad607ed8c9ed0cc0babe6c6a32d