Remove -i and -k flags from make invocations
Abe should not use the -i (--ignore-errors) option for make. This causes
errors which occur during building to be undetected.

This patch also removes -k (--keep-going) as it makes it harder to diagnose
errors from build logs.

Since "make check" is expected to have some failures, we don't want to abort
those when failures occur, so "-i -k" is retained for make check.

Change-Id: Ie23640c752968f052adfdcc92a80bba0e35c8619