Update patch set 2
Patch Set 2:

I had forgotten that DEJAGNU is an optional variable.

I'm OK in principle, but it seems this patch makes the testsuite fail.

I did rebase it to make sure it's applied on top of your recent testsuite fixes, so I think it should pass, and I can't see why this patch would make it fail....

Patch-set: 2
Label: Code-Review=0