Ignore setting a libc version incompatible with the target, unless --set libc=XXX is used.
This allows:
$ abe.sh --target arm-none-eabi glibc=XXX newlib=YYY
or
$ abe.sh --target arm-linux-gnueabi glibc=XXX newlib=YYY
to succeed, using the appropriate libc.

This makes it simpler to implement buildfarm jobs where
we want to build several targets with given C library
versions. Otherwise, we'd have to specialize the
calls to abe depending on the target.

This makes it possible to have a 'backport'-type job
decide which libc branches to use according to the
gcc branch, and to provide buildfarm jobs with these
versions, whatever the target types being built.

Change-Id: Ifd0716ba01374343bf665af522a9af1e9f64e2d3