Remove libc_cv_ctors_header=yes from default configure options
binutils has long been putting all of .dtors and .ctors section
contents into .init_array and .fini_array and as a result, enforcing
libc_cv_ctors_header=yes only results in a table with an invalid
pointer that the dynamic linker tries to execute and fails, resulting
in a segfault.

This does not show up with statically linked binaries, only
dynamically linked ones since the code is not built in the static
case.

Change-Id: I3e415a69f29bfbe7c8c581c0a71888107b8a612b