Update patch set 2
Patch Set 2:

Fair enough, although I doubt if it will cause any issues in gcc5.  I don't see how this can ever work with a dynamic library given that it will always try to execute a function at address (void *)-1.  The only possibilities are that users are either only doing static builds or are ending up not using the sysroot during execution of their programs.  It is very likely the latter.

Patch-set: 2