Update patch set 7
Patch Set 7:

The online documentation needs an update to mention that 'newlib' is an accepted value for --check, --excludecheck, ...

I've checked that newlib tests are actually executed with this patch. However, since we build many multilibs, we end up with many newlib.sum/newlib.log files which are going to confuse jenkins.sh when it saves the results to the log server.

Patch-set: 7
Label: Code-Review=0