Update patch set 6
Patch Set 6:

Rob,

Our convention is that +1 or +2 on a previous version of the patch applies to the subsequent one (e.g., for when the reviewer asked for a simple change or typo-type fixes).

So just +2 yourself and merge.

Patch-set: 6
Reviewer: Maxim Kuvyrkov <1000092@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0