Update patch set 2
Patch Set 2:

Siddhesh,

Our builds support testing using QEMU, and on an x86_64 host there is definitely only our just-built armhf / arm64 sysroot to use, and we see tests executing fine.

You can reproduce this with testing using ABE master with --disable-schroot-test.

In any case, it is a good idea for us to test in a schroot where system libs are incompatible with the testcase executables.  E.g., test armhf in an arm64 sysroot and vice versa, to make sure just-built sysroots are used.

Patch-set: 2