Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

The real problem is that checkout does not handle gdbserver correctly by treating it as a special case which means that the gdbserver revision is not set in the component info, and the gdbserver sources are not checked out.

I think the right fix is to remove this code from checkout()
    # gdbserver is already checked out in the GDB source tree.
    if test x"${component}" = x"gdbserver"; then
        return 0
    fi

[ this should also fix build failure with this manifest https://pastebin.linaro.org/view/fd71532b ]

We should also bump the manifest version number to 1.1, and reject 1.0 manifests, since they are useless due to the missing *_revision fields. (abe-tests will fail, because it expects 1.1 to have *_md5sums, I will correct that once a patch is available)

Patch-set: 1
Reviewer: Charles Baylis <1000108@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=-1