Update patch set 2
Patch Set 2:

The difference between native and cross is in the linker script.  The native linker has
  .init_array     :
  {
    PROVIDE_HIDDEN (__init_array_start = .);
    KEEP (*(SORT_BY_INIT_PRIORITY(.init_array.*) SORT_BY_INIT_PRIORITY(.ctors.*)))
    KEEP (*(.init_array EXCLUDE_FILE (*crtbegin.o *crtbegin?.o *crtend.o *crtend?.o ) .ctors))
    PROVIDE_HIDDEN (__init_array_end = .);
  }
the cross linker has
  .init_array     :
  {
    PROVIDE_HIDDEN (__init_array_start = .);
    KEEP (*(SORT(.init_array.*)))
    KEEP (*(.init_array ))
    PROVIDE_HIDDEN (__init_array_end = .);
  }
So the native linker puts ctor sections in init_array and the cross doesn't.

This difference is due to enable_initfini_array, which gets set for the native at configure time and is not set for the cross.  This happens because there is a run-time test in the configure script, which doesn't work cross, so this is disabled for cross and enabled for native.  This run-time test was removed here

2015-12-02  Alan Modra  <amodra@gmail.com>

        * configure.ac (--enable-initfini-array): Remove run test.  Default
        to "yes".  Change help string to --disable-initfini-array.
        * configure: Regenerate.

To make the native and cross linkers work the same, we either need to backport this patch, or add an explicit --enable-initfini-array configure option, or add an explicit --disable-initfini-array configure option.

If we make either of the first two changes, then the cross toolchain will fail same as the native toolchain, and we will need Siddhesh's patch to remove libc_cv_ctors_header=yes from glibc configure to make them both work.

If we make the third change, then native will work same as cross, and Siddhesh's patch is not needed, but long term we should not be disabling the initfini support, as this is how everything should work in the future.

Patch-set: 2
Reviewer: Jim Wilson <1000132@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0