Move version configuration for gcc5-based toolchains to a dedicated dir
This patch adds logic to source .conf snipets from config/default/ directory,
which is a symlink to config/gcc5/ directory.  The .conf snipets define
versions of components we are currently using.

Note that configuration in config/default/ files should be limited to setting
"latest=" field for components, and the main configuration logic should be in
config/*.conf files.  It is OK, however, to have extra configuration logic in
non-default config/gccN directories.

An upcoming patch will add config/gcc6 directory for gcc6-based toolchains,
and we will then switch the default to gcc6 just by updating the symlink.

At that point we will be still able to easily build gcc5-based toolchains by
adding --extraconfigdir config/gcc5 to a build.

Change-Id: Icc99ee0624ba56c3f8573784e741ba344110b410