Update patch set 2
Patch Set 2:

The current design doesn't ensure that gdb and gdbserver are built from the same sources. It simply doesn't check out gdbserver and hopes that the checkout of gdb will provide gdbserver sources in the correct directory. Try this manifest to see that it doesn't always work https://pastebin.linaro.org/view/fd71532b.

Since gdbserver and gdb are closely related, it could make sense to tie their revisions together, but that should be done by ensuring that the component array is filled in appropriately. This would mean that only one special case is required.

Patch-set: 2