Fix bug #2287 and fix gdbserver special case in checkout().
Reinstate the commented out call to set_component_revision, which fixes the
*_revision field for most components.

For gdbserver, take revision from gdb checkout. Return an error if gdb and
gdbserver are to be built from different source trees.

This patch also bumps the manifest version number to 1.1 so that useless
manifests without the *_revision fields are rejected as invalid.

Change-Id: Ic511f42e50219fb2639a816e97d44ba5452415c2