Fix bug #2287 and removing gdbserver special case in checkout().
Reinstate the commented out call to set_component_revision, which fixes the
*_revision field for most components.

Remove the special case in checkout() for gdbserver so that the component array
is populated with correct revision information for gdbserver, thereby causing
the manifest to contain gdbserver_revision.  This special case was intended to
ensure that gdbserver is built from the same sources as gdb, but is not
actually effective, and in fact gdbserver can readily be built from different
sources if the manifest contains different revisions for gdb and gdbserver, or
if the user uses gdb=binutils-gdb.git~foo on the abe command line.

This patch also bumps the manifest version number to 1.1 so that useless
manifests without the *_revision fields are rejected as invalid.

Change-Id: Ic511f42e50219fb2639a816e97d44ba5452415c2