Update patch set 1
Patch Set 1:

This isn't a matter of "future enhancements", it is about a bug which is in this patch. It is a matter of error handling.

With this patch, the checkout() function makes an assumption that gdb and gdbserver, and if that assumption is false then it silently produces an incorrect manifest.

If for some reason, "get_component_srcdir gdb" and "get_component_srcdir gdbserver" don't point to the same directory, then this patch results does "set_component_revision gdbserver" with a revision that doesn't necessarily match the gdbserver source tree.

The fact that gdb and gdbserver "should" be built from the same source is irrelevant. There is a serious failure mode in this implementation of checkout(). This would warrant a -1 even if it was only a latent bug.

However, in reality, this bug is not latent, and can easily be triggered with the right abe command line. This demonstrates that:
. this implementation of checkout() is unsafe
. it is difficult to correctly implement the appropriate checks
. this patch has genuine bugs.

There is a demonstration of some of these bugs in abe-tests/gdbserver-mismatch.sh. Run with ABE_PATH and GIT_REF_DIR environment variables set appropriately.

Patch-set: 1