Update patch set 1
Patch Set 1:

GDBSERVER is a sub-directory of GDB, that's why the srcdir paths are different. The main difference between the two is GDB gets built with the build machine gcc, and GDBSERVER gets built for host/target. They should both *always* be built from the same source revision. Breaking ABE via a hack isn't appropriate. This fixes a critical bug needed for this release. Any support for building GDB and and GDBSERVER from different revisions would only be used by a tiny handful of GDB developers, and is *not* recommended. While this coud be implemented in ABE, it's a new feature, and not covered by this bug, which was to put the revision fields back in the manifest file.

Patch-set: 1
Reviewer: Rob Savoye <1000040@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0