Update patch set 2
Patch Set 2:

Rob, this change doesn't make it checkout, but it uses the arguments to construct a revision for gdbserver that matches that of GDB, and makes sure they can't be different.

In that sense, there is no new behaviour at all, only a stronger check to see that the versions are indeed the same.

If you just skip the whole process and assume they are the same, and they're not, then you'll have trouble. In this case, such a problem cannot happen.

Patch-set: 2
Reviewer: Renato Golin <1000088@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0