Update patch set 3
Patch Set 3:

(1 comment)

This looks fine to me, but splitting the default configuration over 2 files will make it easier to make a mistake when changing the default.

Since the conf files are mostly free-form, it is difficult to make a good sanity check. But something is probably better than nothing.

Other than that, LGTM.

Patch-set: 3