Update patch set 3
Patch Set 3:

Charles,

I've also thought about the switch scenario that you describe.  I agree that the use of "config/default" symlink is a very restrictive solution in terms of features it provides.  I thought of implementing a --pre-config{-dir}= and --post-config{-dir}= options, but it seems as an overkill for the functionality we currently need, and I'm not sure we will use these options if they are implemented.

I do agree that the symlink solution can be fragile and changes to "config/gccN" need to be audited properly.  I still think this patch is an easy and simple first step to get functionality that we need.

Patch-set: 3