Use wget -c instead of -N to workaround s.l.o 403 response.
With -N, earlier versions of wget (<=1.16) use a HEAD request to determine
whether the server has a newer version of the file. Unfortunately, s.l.o now
returns 403 Forbidden for HEAD requests, so this fails.

Wget >= 1.17 uses the If-Modified-Since header in a GET request instead, but
we can't assume that all our users will upgrade.

Therefore, this patch uses the -c option to wget. This will cause wget to
complete the download of an existing file, or return successfully if the file
is already fully downloaded.

Change-Id: I679a30bd1e3b5787b6ffcb08f89da24e67227ad3